Add interface to perspective integration
